Now it works correctly with VEManager, fun-bots and other mods such as compass.
Implemented the automatic switch of the nvg when we get in or out of a vehicle.

Added the thermal vision to the vehicle nvg preset.

Added config options for vehicle lights.

Removed the patchAssets and the patching we were doing for meshes and other stuff that seems unnecesserary now that VEM already does that without issues.

TODO: Add the vehicle lights but make them toggleable and default them to be off.