Updated dependencies.
Refactor and cleanup

Small changes in the NVG and Vehicle NVG presets.

Removed the PatchAssets script since we don't need it anymore and it was causing issues with VEM since its already patching most of the stuff.

Kept the patches script since we are still doing some things that VEM doesn't and maybe it shouldn't at all, for example, the bullet, impacts, explossions emitters, flashlights and such.

Added some UI stuff, a cute animation for the situation where the NVG is depleted so it bounces the nvg goggle icon and some red color.
Also changed some audio files for the NVG.

TODO: I'd like to be able to "pulse" or make some kind of glitch preset for a couple seconds if the player tries to turn on the NVG while still on cooldown (Tried using Pulse from VEM but it wasn't working properly)

Config file cleanup, the time settings at least for now. Enabling time as soon as the server can will cause the game to crash as soon as a player gets in, still don't know the cause of this.